// src/app/agents/interfaces/agent.ts
export class Agent {
  Name: string;
  Addr1: string;
  City: string;
  Postcode: number;
  State: string;
  ContactName: string;
  URL: string;
  Country: string;
  ParentAgentName: string;
  ParentAgentContact: string;
  ParentAgentStreet: string;
  ParentAgentCity: string;
  ParentAgentState: string;
  ParentAgentPostcode: string;
  ParentAgentCountry: string;
  ParentAgentWebsite: string;
}